+++ docs/js/menu-wc.js
@@ -475,13 +475,13 @@ customElements.define('compodoc-menu', class extends HTMLElement {
                                 <a href="modules/FormModule.html" data-type="entity-link">FormModule</a>
                                     <li class="chapter inner">
                                         <div class="simple menu-toggler" data-toggle="collapse" ${ isNormalMode ?
-                                            'data-target="#components-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' : 'data-target="#xs-components-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' }>
+                                            'data-target="#components-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' : 'data-target="#xs-components-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' }>
                                             <span class="icon ion-md-cog"></span>
                                             <span>Components</span>
                                             <span class="icon ion-ios-arrow-down"></span>
                                         </div>
-                                        <ul class="links collapse" ${ isNormalMode ? 'id="components-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' :
-                                            'id="xs-components-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' }>
+                                        <ul class="links collapse" ${ isNormalMode ? 'id="components-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' :
+                                            'id="xs-components-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' }>
                                             <li class="link">
                                                 <a href="components/DatetimeComponent.html"
                                                     data-type="entity-link" data-context="sub-entity" data-context-id="modules">DatetimeComponent</a>
@@ -506,13 +506,13 @@ customElements.define('compodoc-menu', class extends HTMLElement {
                                     </li>
                                 <li class="chapter inner">
                                     <div class="simple menu-toggler" data-toggle="collapse" ${ isNormalMode ?
-                                        'data-target="#injectables-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' : 'data-target="#xs-injectables-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' }>
+                                        'data-target="#injectables-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' : 'data-target="#xs-injectables-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' }>
                                         <span class="icon ion-md-arrow-round-down"></span>
                                         <span>Injectables</span>
                                         <span class="icon ion-ios-arrow-down"></span>
                                     </div>
-                                    <ul class="links collapse" ${ isNormalMode ? 'id="injectables-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' :
-                                        'id="xs-injectables-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' }>
+                                    <ul class="links collapse" ${ isNormalMode ? 'id="injectables-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' :
+                                        'id="xs-injectables-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' }>
                                         <li class="link">
                                             <a href="injectables/FormService.html"
                                                 data-type="entity-link" data-context="sub-entity" data-context-id="modules" }>FormService</a>
@@ -521,13 +521,13 @@ customElements.define('compodoc-menu', class extends HTMLElement {
                                 </li>
                                     <li class="chapter inner">
                                         <div class="simple menu-toggler" data-toggle="collapse" ${ isNormalMode ?
-                                            'data-target="#pipes-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' : 'data-target="#xs-pipes-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' }>
+                                            'data-target="#pipes-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' : 'data-target="#xs-pipes-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' }>
                                             <span class="icon ion-md-add"></span>
                                             <span>Pipes</span>
                                             <span class="icon ion-ios-arrow-down"></span>
                                         </div>
-                                        <ul class="links collapse" ${ isNormalMode ? 'id="pipes-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' :
-                                            'id="xs-pipes-links-module-FormModule-14eb46782c004a28d99db16cf2c29bae"' }>
+                                        <ul class="links collapse" ${ isNormalMode ? 'id="pipes-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' :
+                                            'id="xs-pipes-links-module-FormModule-c8ba2c78dd4c9c4ada232b8cc4e56c79"' }>
                                             <li class="link">
                                                 <a href="pipes/MaxItemsPipe.html"
                                                     data-type="entity-link" data-context="sub-entity" data-context-id="modules">MaxItemsPipe</a>